datagen/requirements.txt
python-kafka